fix(connect): refuse a changed operation before anything can publish one, and share the mutation lock
Four review findings, and the lower half of the exclusion primitive.

The operation check a resumed run makes was reached *after* the prepare step,
which creates a draft when none exists — so a run whose operation had vanished
answered by publishing a new one and then rejecting it, leaving a phantom pending
operation for the next run to resume. The question is now settled before anything
can create: same operation proceeds, a different one is refused, and a vanished
one is refused with the two causes distinguished, because "another process
finished enrolling this directory" and "the operation was removed" need different
next steps. Nothing is sent and no key is redeemed in either refusal.

The two filesystem probes in the non-interactive guard were synchronous calls on
an async path, and `Path::try_exists` for the draft sat beside `Path::exists` for
the identity; both now go through `tokio::fs`.

`IdentityStore::set_attachment` stored the Cloud-delivered monitor page unchecked,
and the runtime prints it — the same exposure the create-project page had, at a
different writer. It passes the same portal-link rule there, so a rejected link
costs the page and not the attachment.

A pods-watcher log line embedded a newline, which the logging rule forbids.

`ConnectLock` moves out of the CLI into `runtime_cloud_connect::mutation_lock` as
`MutationLock`, keeping `connect.lock` so it is the same lock file — the runtime
reaches this state through the control stream and has to take the same lock, and
two exclusion primitives over one directory would exclude nothing. The module
states its order relative to `EnrollmentTransactionLock`: this one is outer, taken
first. Its four tests prove the semantics where they now live — one holder at a
time with the waiter told which action and pid holds it, bounded diagnostics
stripped of control characters, a planted symlink never becoming the lock, and a
symlinked ancestor resolving to one lock rather than two.
